+++ concordium-consensus/tests/scheduler/SchedulerTests/SmartContracts/V1/Queries.hs
@@ -4,16 +4,14 @@
 module SchedulerTests.SmartContracts.V1.Queries (tests) where
 
 import Test.Hspec
-import Test.HUnit(assertFailure, assertEqual, assertBool)
+import Test.HUnit(assertFailure, assertEqual)
 
 import Lens.Micro.Platform
 import qualified Data.ByteString.Short as BSS
 import qualified Data.ByteString as BS
 import Control.Monad
-import qualified Data.Set as Set
-import Data.Serialize(runPut, Serialize (put), putWord32le, putWord64be)
+import Data.Serialize(runPut, Serialize (put), putWord64le)
 import qualified Data.Text as T
-import qualified Data.List as List
 
 import qualified Concordium.Scheduler.Types as Types
 import qualified Concordium.Crypto.SHA256 as Hash
@@ -22,7 +20,6 @@ import qualified Concordium.TransactionVerification as TVer
 
 import Concordium.GlobalState.Basic.BlockState.Accounts as Acc
 import Concordium.GlobalState.Basic.BlockState
-import qualified Concordium.GlobalState.Wasm as GSWasm
 import Concordium.Wasm
 import qualified Concordium.Cost as Cost
 
@@ -33,6 +30,7 @@ import Concordium.Crypto.DummyData
 
 import SchedulerTests.TestUtils
 import Concordium.Types.UpdateQueues (currentParameters)
+import qualified Concordium.Wasm as Wasm
 
 thomasBalance :: Types.Amount
 thomasBalance = 100000000
@@ -45,7 +43,6 @@ initialBlockState = blockStateWithAlesAccount
 blockEnergyRate :: Types.EnergyRate
 blockEnergyRate = initialBlockState ^. blockUpdates . currentParameters . Types.energyRate
 
--- |A module that is used as a base for upgrading.
 accountBalanceSourceFile :: FilePath
 accountBalanceSourceFile = "./testdata/contracts/v1/queries-account-balance.wasm"
 
@@ -74,9 +71,9 @@ accountBalanceTestCase =
   where
     parameters = BSS.toShort $ runPut $ do
       put thomasAccount
-      put thomasBalance -- expected public balance
-      put (0 :: Types.Amount) -- expected staked amount
-      put (0 :: Types.Amount) -- expected locked amount
+      putWord64le (fromIntegral thomasBalance) -- expected public balance
+      putWord64le 0
+      putWord64le 0
 
     eventsCheck :: [Types.Event] -> Expectation
     eventsCheck events = do
@@ -112,9 +109,9 @@ accountBalanceInvokerTestCase =
 
     parameters = BSS.toShort $ runPut $ do
       put thomasAccount
-      put $ thomasBalance - costUpperBound - updateAmount -- expected public balance
-      put (0 :: Types.Amount) -- expected staked amount
-      put (0 :: Types.Amount)  -- expected locked amount
+      Wasm.putAmountLE $ thomasBalance - costUpperBound - updateAmount -- expected public balance
+      Wasm.putAmountLE 0 -- expected staked amount
+      Wasm.putAmountLE 0 -- expected locked amount
 
     costUpperBound = Types.computeCost blockEnergyRate energyLimit
 
@@ -123,6 +120,296 @@ accountBalanceInvokerTestCase =
         -- Check the number of events:
         -- - 1 event for a succesful update to the contract ('contract.upgrade').
         eventsLengthCheck 1 events
+
+accountBalanceTransferSourceFile :: FilePath
+accountBalanceTransferSourceFile = "./testdata/contracts/v1/queries-account-balance-transfer.wasm"
+
+accountBalanceTransferTestCase :: TestCase PV5
+accountBalanceTransferTestCase =
+  TestCase { tcName = "Contracts transfers to an account and then queries"
+           , tcParameters = (defaultParams @PV5) {tpInitialBlockState=initialBlockState}
+           , tcTransactions =
+             [ ( TJSON { payload = DeployModule V1 accountBalanceTransferSourceFile
+                       , metadata = makeDummyHeader alesAccount 1 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (deploymentCostCheck accountBalanceTransferSourceFile), emptySpec))
+             , ( TJSON { payload = InitContract 123 V1 accountBalanceTransferSourceFile "init_contract" ""
+                       , metadata = makeDummyHeader alesAccount 2 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (initializationCostCheck accountBalanceTransferSourceFile "init_contract"), emptySpec))
+             , ( TJSON { payload = Update 0 (Types.ContractAddress 0 0) "contract.query" parameters
+                       , metadata = makeDummyHeader alesAccount 3 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (successWithEventsCheck eventsCheck), emptySpec))
+             ]
+           }
+  where
+    parameters = BSS.toShort $ runPut $ do
+      put thomasAccount
+      Wasm.putAmountLE 123
+      Wasm.putAmountLE $ thomasBalance + 123 -- expected public balance
+      Wasm.putAmountLE 0 -- expected staked amount
+      Wasm.putAmountLE 0 -- expected locked amount
+
+    eventsCheck :: [Types.Event] -> Expectation
+    eventsCheck events = do
+        -- Check the number of events:
+        -- - 3 events for the transfer.
+        -- - 1 event for a succesful update to the contract.
+        eventsLengthCheck 4 events
+
+accountBalanceMissingAccountSourceFile :: FilePath
+accountBalanceMissingAccountSourceFile = "./testdata/contracts/v1/queries-account-balance-missing-account.wasm"
+
+accountBalanceMissingAccountTestCase :: TestCase PV5
+accountBalanceMissingAccountTestCase =
+  TestCase { tcName = "Query the balance of a missing account"
+           , tcParameters = (defaultParams @PV5) {tpInitialBlockState=initialBlockState}
+           , tcTransactions =
+             [ ( TJSON { payload = DeployModule V1 accountBalanceMissingAccountSourceFile
+                       , metadata = makeDummyHeader alesAccount 1 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (deploymentCostCheck accountBalanceMissingAccountSourceFile), emptySpec))
+             , ( TJSON { payload = InitContract 123 V1 accountBalanceMissingAccountSourceFile "init_contract" ""
+                       , metadata = makeDummyHeader alesAccount 2 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (initializationCostCheck accountBalanceMissingAccountSourceFile "init_contract"), emptySpec))
+             , ( TJSON { payload = Update 0 (Types.ContractAddress 0 0) "contract.query" parameters
+                       , metadata = makeDummyHeader alesAccount 3 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (successWithEventsCheck eventsCheck), emptySpec))
+             ]
+           }
+  where
+    parameters = BSS.toShort $ runPut $ do
+      put $ accountAddressFrom 3
+
+    eventsCheck :: [Types.Event] -> Expectation
+    eventsCheck events = do
+        -- Check the number of events:
+        -- - 1 event for a succesful update to the contract.
+        eventsLengthCheck 1 events
+
+contractBalanceSourceFile :: FilePath
+contractBalanceSourceFile = "./testdata/contracts/v1/queries-contract-balance.wasm"
+
+contractBalanceTestCase :: TestCase PV5
+contractBalanceTestCase =
+  TestCase { tcName = "Query the balance a contract"
+           , tcParameters = (defaultParams @PV5) {tpInitialBlockState=initialBlockState}
+           , tcTransactions =
+             [ ( TJSON { payload = DeployModule V1 contractBalanceSourceFile
+                       , metadata = makeDummyHeader alesAccount 1 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (deploymentCostCheck contractBalanceSourceFile), emptySpec))
+             , ( TJSON { payload = InitContract initAmount V1 contractBalanceSourceFile "init_contract" ""
+                       , metadata = makeDummyHeader alesAccount 2 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (initializationCostCheck contractBalanceSourceFile "init_contract"), emptySpec))
+             , ( TJSON { payload = Update updateAmount (Types.ContractAddress 0 0) "contract.query" parameters
+                       , metadata = makeDummyHeader alesAccount 3 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (successWithEventsCheck eventsCheck), emptySpec))
+             ]
+           }
+  where
+    initAmount = 123
+    updateAmount = 456
+
+    parameters = BSS.toShort $ runPut $ do
+      putWord64le 0 -- Index of the contract address.
+      putWord64le 0 -- Subindex of the contract address.
+      Wasm.putAmountLE $ initAmount + updateAmount -- Expected balance.
+
+    eventsCheck :: [Types.Event] -> Expectation
+    eventsCheck events = do
+        -- Check the number of events:
+        -- - 1 event for a succesful update to the contract.
+        eventsLengthCheck 1 events
+
+contractBalanceTransferSourceFile :: FilePath
+contractBalanceTransferSourceFile = "./testdata/contracts/v1/queries-contract-balance-transfer.wasm"
+
+contractBalanceTransferTestCase :: TestCase PV5
+contractBalanceTransferTestCase =
+  TestCase { tcName = "Query the balance a contract"
+           , tcParameters = (defaultParams @PV5) {tpInitialBlockState=initialBlockState}
+           , tcTransactions =
+             [ ( TJSON { payload = DeployModule V1 contractBalanceTransferSourceFile
+                       , metadata = makeDummyHeader alesAccount 1 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (deploymentCostCheck contractBalanceTransferSourceFile), emptySpec))
+             , ( TJSON { payload = InitContract initAmount V1 contractBalanceTransferSourceFile "init_contract" ""
+                       , metadata = makeDummyHeader alesAccount 2 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (initializationCostCheck contractBalanceTransferSourceFile "init_contract"), emptySpec))
+             , ( TJSON { payload = Update updateAmount (Types.ContractAddress 0 0) "contract.query" parameters
+                       , metadata = makeDummyHeader alesAccount 3 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (successWithEventsCheck eventsCheck), emptySpec))
+             ]
+           }
+  where
+    initAmount = 123
+    updateAmount = 456
+    transferAmount = 78
+
+    parameters = BSS.toShort $ runPut $ do
+      put thomasAccount
+      Wasm.putAmountLE transferAmount
+      putWord64le 0 -- Index of the contract address.
+      putWord64le 0 -- Subindex of the contract address.
+      Wasm.putAmountLE $ initAmount + updateAmount - transferAmount -- Expected balance.
+
+    eventsCheck :: [Types.Event] -> Expectation
+    eventsCheck events = do
+        -- Check the number of events:
+        -- - 3 events for transfering
+        -- - 1 event for a succesful update to the contract.
+        eventsLengthCheck 4 events
+
+contractBalanceMissingContractSourceFile :: FilePath
+contractBalanceMissingContractSourceFile = "./testdata/contracts/v1/queries-contract-balance-missing-contract.wasm"
+
+contractBalanceMissingContractTestCase :: TestCase PV5
+contractBalanceMissingContractTestCase =
+  TestCase { tcName = "Query the balance of a missing contract"
+           , tcParameters = (defaultParams @PV5) {tpInitialBlockState=initialBlockState}
+           , tcTransactions =
+             [ ( TJSON { payload = DeployModule V1 contractBalanceMissingContractSourceFile
+                       , metadata = makeDummyHeader alesAccount 1 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (deploymentCostCheck contractBalanceMissingContractSourceFile), emptySpec))
+             , ( TJSON { payload = InitContract 123 V1 contractBalanceMissingContractSourceFile "init_contract" ""
+                       , metadata = makeDummyHeader alesAccount 2 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (initializationCostCheck contractBalanceMissingContractSourceFile "init_contract"), emptySpec))
+             , ( TJSON { payload = Update 0 (Types.ContractAddress 0 0) "contract.query" parameters
+                       , metadata = makeDummyHeader alesAccount 3 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (successWithEventsCheck eventsCheck), emptySpec))
+             ]
+           }
+  where
+    parameters = BSS.toShort $ runPut $ do
+      putWord64le 1 -- Contract address index
+      putWord64le 0 -- Contract address subindex
+
+    eventsCheck :: [Types.Event] -> Expectation
+    eventsCheck events = do
+        -- Check the number of events:
+        -- - 1 event for a succesful update to the contract.
+        eventsLengthCheck 1 events
+
+exchangeRatesSourceFile :: FilePath
+exchangeRatesSourceFile = "./testdata/contracts/v1/queries-exchange-rates.wasm"
+
+exchangeRatesTestCase :: TestCase PV5
+exchangeRatesTestCase =
+  TestCase { tcName = "Query the exchange rates"
+           , tcParameters = (defaultParams @PV5) {tpInitialBlockState=initialBlockState}
+           , tcTransactions =
+             [ ( TJSON { payload = DeployModule V1 exchangeRatesSourceFile
+                       , metadata = makeDummyHeader alesAccount 1 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (deploymentCostCheck exchangeRatesSourceFile), emptySpec))
+             , ( TJSON { payload = InitContract 0 V1 exchangeRatesSourceFile "init_contract" ""
+                       , metadata = makeDummyHeader alesAccount 2 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (initializationCostCheck exchangeRatesSourceFile "init_contract"), emptySpec))
+             , ( TJSON { payload = Update 0 (Types.ContractAddress 0 0) "contract.query" parameters
+                       , metadata = makeDummyHeader alesAccount 3 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (successWithEventsCheck eventsCheck), emptySpec))
+             ]
+           }
+  where
+    parameters = BSS.toShort $ runPut $ do
+      Wasm.putExchangeRateLE currentEuroPerEnergy
+      Wasm.putExchangeRateLE currentAmountPerEnergy
+
+    currentEuroPerEnergy = initialBlockState ^. blockUpdates . currentParameters . Types.euroPerEnergy
+    currentAmountPerEnergy = initialBlockState ^. blockUpdates . currentParameters . Types.microGTUPerEuro
+
+    eventsCheck :: [Types.Event] -> Expectation
+    eventsCheck events = do
+        -- Check the number of events:
+        -- - 1 event for a succesful update to the contract.
+        eventsLengthCheck 1 events
+
+allSourceFile :: FilePath
+allSourceFile = "./testdata/contracts/v1/queries-all.wasm"
+
+allTestCase :: TestCase PV4
+allTestCase =
+  TestCase { tcName = "Ensure all of the queries fail prior to PV5"
+           , tcParameters = (defaultParams @PV4) {tpInitialBlockState=initialBlockStateP4}
+           , tcTransactions =
+             [ ( TJSON { payload = DeployModule V1 allSourceFile
+                       , metadata = makeDummyHeader alesAccount 1 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (deploymentCostCheck allSourceFile), emptySpec))
+             , ( TJSON { payload = InitContract 0 V1 allSourceFile "init_contract" ""
+                       , metadata = makeDummyHeader alesAccount 2 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (initializationCostCheck allSourceFile "init_contract"), emptySpec))
+
+             , ( TJSON { payload = Update 0 (Types.ContractAddress 0 0) "contract.account_balance" ""
+                       , metadata = makeDummyHeader alesAccount 3 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (rejectWithReasonCheck rejectReasonCheck), emptySpec))
+
+               , ( TJSON { payload = Update 0 (Types.ContractAddress 0 0) "contract.contract_balance" ""
+                       , metadata = makeDummyHeader alesAccount 4 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (rejectWithReasonCheck rejectReasonCheck), emptySpec))
+
+               , ( TJSON { payload = Update 0 (Types.ContractAddress 0 0) "contract.exchange_rates" ""
+                       , metadata = makeDummyHeader alesAccount 5 100000
+                       , keys = [(0,[(0, alesKP)])]
+                       }
+               , (SuccessWithSummary (rejectWithReasonCheck rejectReasonCheck), emptySpec))
+             ]
+           }
+  where
+    eventsCheck :: [Types.Event] -> Expectation
+    eventsCheck events = do
+        -- Check the number of events:
+        -- - 1 event for a succesful update to the contract.
+        eventsLengthCheck 1 events
+
+    initialBlockStateP4 :: BlockState PV4
+    initialBlockStateP4 = blockStateWithAlesAccount
+      100000000
+      (Acc.putAccountWithRegIds (mkAccount thomasVK thomasAccount thomasBalance) Acc.emptyAccounts)
+
+    rejectReasonCheck :: Types.RejectReason -> Expectation
+    rejectReasonCheck reason =
+      case reason of
+          Types.RuntimeFailure -> return ()
+          other -> assertFailure $ "Unexpected reject reason " ++ show other
 
 -- This only checks that the cost of initialization is correct.
 -- If the state was not set up correctly the latter tests in the suite will fail.
@@ -168,20 +455,26 @@ successWithEventsCheck checkEvents _ summary = case Types.tsResult summary of
   Types.TxReject {..} -> assertFailure $ "Transaction rejected unexpectedly with " ++ show vrRejectReason
   Types.TxSuccess {..} -> checkEvents vrEvents
 
+-- | Check the number of events is as expected.
+eventsLengthCheck :: Int -> [Types.Event] -> Expectation
+eventsLengthCheck expected events = unless (length events == expected) $
+  assertFailure $ "Unexpected number of events produced: " ++ show (length events) ++ " where the expected was " ++ show expected
+
 -- | Check the transaction rejected, taking a function to check the reason.
 rejectWithReasonCheck :: (Types.RejectReason -> Expectation) -> TVer.BlockItemWithStatus -> Types.TransactionSummary -> Expectation
 rejectWithReasonCheck checkReason _ summary = case Types.tsResult summary of
   Types.TxReject {..} -> checkReason vrRejectReason
   Types.TxSuccess {} -> assertFailure "Transaction succeeded unexpectedly"
 
--- | Check the number of events is as expected.
-eventsLengthCheck :: Int -> [Types.Event] -> Expectation
-eventsLengthCheck expected events = unless (length events == expected) $
-  assertFailure $ "Unexpected number of events produced: " ++ show (length events) ++ " where the expected was " ++ show expected
-
-
 tests :: Spec
-tests = describe "V1: Queries" $ mkSpecs [
+tests = describe "V1: Queries" $ do
+  mkSpecs [
     accountBalanceTestCase,
-    accountBalanceInvokerTestCase
-  ]
+    accountBalanceInvokerTestCase,
+    accountBalanceTransferTestCase,
+    accountBalanceMissingAccountTestCase,
+    contractBalanceTestCase,
+    contractBalanceTransferTestCase,
+    contractBalanceMissingContractTestCase,
+    exchangeRatesTestCase]
+  mkSpecs [allTestCase]