+++ concordium-consensus/tests/scheduler/SchedulerTests/SmartContracts/V1/RelaxedRestrictions.hs
@@ -11,7 +11,7 @@
         - Of size <= 1kb: base cost + 1NRG / 1 *kilobyte* (same as before P5)
         - Of size > 1 kb: base cost + 1NRG / 1 *byte*
 -}
-module SchedulerTests.SmartContracts.V1.RelaxedRestrictions (tests) where
+module SchedulerTests.SmartContracts.V1.RelaxedRestrictions (testsPV4, testsPV5) where
 
 import Test.Hspec
 import Test.HUnit(assertFailure, assertEqual)
@@ -316,9 +316,10 @@ checkSuccess :: [Char] -> Types.ValidResult -> IO ()
 checkSuccess msg Types.TxReject{..} = assertFailure $ msg ++ show vrRejectReason
 checkSuccess _ _ = return ()
 
-tests :: Spec
-tests = describe "V1: Relax restrictions." $
-  --[ TODO: Run both tests. It complains about something with the PV.
+testsPV4 :: Spec
+testsPV4 = describe "V1: Relax restrictions. Test in PV4." $
     mkSpecs testCasesPV4
-  --, mkSpecs testCasesPV5
-  --]
+
+testsPV5 :: Spec
+testsPV5 = describe "V1: Relax restrictions. Test in PV5." $
+    mkSpecs testCasesPV5