+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for protocol version 5. This adds the following features
   - support for smart contract upgradability
+  - Query the current exchange rates, account balances and contract balances from a smart contract.
   - Relax restrictions on smart contracts
     - Parameter size limit: 1kb -> 65kb
     - Return value size limit: 16kb -> no limit (apart from energy)